Convergence module Simplifies ToGraph module, splitting it in two modules
